Jupyter Notebook for Analysis of Twitter Trends
This file consists of the Jupyter notebook file for doing the analysis of the twitter trends. You can easily check this code out and try on your own for looking for twitter trends. We used the Twitter API for the same. You just have to make an app in twitter. from where you will get the credentials for the same. And which you have to write in the places as in the notebook.